requirements.txt
boto3==1.34.23
pyyaml==5.3